# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Classroom management backend (Go net/http ServeMux). Auth via JWT Bearer (HS256, 24h); mini-program identity resolved server-side via code2Session.",
        "title": "OCM 智慧教室管理平台 API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/schedule/active": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schedule"
                ],
                "summary": "Get the regime active today",
                "responses": {
                    "200": {
                        "description": "regime with periods",
                        "schema": {
                            "$ref": "#/definitions/schedule.Regime"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/schedule/regimes": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schedule"
                ],
                "summary": "List bell-time regimes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "search by name",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "1-based page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "paged regimes with periods",
                        "schema": {
                            "$ref": "#/definitions/httpx.Paged"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schedule"
                ],
                "summary": "Create a bell-time regime",
                "parameters": [
                    {
                        "description": "regime input",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schedule.RegimeInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "created regime (no periods yet)",
                        "schema": {
                            "$ref": "#/definitions/schedule.Regime"
                        }
                    },
                    "400": {
                        "description": "invalid body / name required / effective date invalid",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "effective date already taken",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/schedule/regimes/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schedule"
                ],
                "summary": "Get a bell-time regime",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "regime id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "regime with periods",
                        "schema": {
                            "$ref": "#/definitions/schedule.Regime"
                        }
                    },
                    "400": {
                        "description": "invalid regime id",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "regime not found",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schedule"
                ],
                "summary": "Update a bell-time regime",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "regime id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "regime input",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schedule.RegimeInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "updated regime with periods",
                        "schema": {
                            "$ref": "#/definitions/schedule.Regime"
                        }
                    },
                    "400": {
                        "description": "invalid body / name required / effective date invalid",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "regime not found",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "effective date already taken",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "schedule"
                ],
                "summary": "Delete a bell-time regime",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "regime id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "no content"
                    },
                    "400": {
                        "description": "invalid regime id",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "regime not found",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "cannot delete the last regime",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/schedule/regimes/{id}/periods": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schedule"
                ],
                "summary": "Replace a regime's periods (bell times)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "regime id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "full replacement set of periods",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schedule.PeriodsInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "regime with the new periods",
                        "schema": {
                            "$ref": "#/definitions/schedule.Regime"
                        }
                    },
                    "400": {
                        "description": "invalid body / period numbers must be contiguous",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "regime not found",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "httpx.ErrorResponse": {
            "type": "object",
            "required": [
                "error"
            ],
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid request body"
                }
            }
        },
        "httpx.Paged": {
            "type": "object",
            "required": [
                "items",
                "page",
                "pageSize",
                "total"
            ],
            "properties": {
                "items": {},
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "schedule.Period": {
            "type": "object",
            "required": [
                "endTime",
                "id",
                "periodIndex",
                "regimeId",
                "startTime"
            ],
            "properties": {
                "endTime": {
                    "description": "\"HH:MM\"",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "periodIndex": {
                    "type": "integer"
                },
                "regimeId": {
                    "type": "integer"
                },
                "startTime": {
                    "description": "\"HH:MM\"",
                    "type": "string"
                }
            }
        },
        "schedule.PeriodInput": {
            "type": "object",
            "required": [
                "endTime",
                "periodIndex",
                "startTime"
            ],
            "properties": {
                "endTime": {
                    "type": "string"
                },
                "periodIndex": {
                    "type": "integer"
                },
                "startTime": {
                    "type": "string"
                }
            }
        },
        "schedule.PeriodsInput": {
            "type": "object",
            "required": [
                "periods"
            ],
            "properties": {
                "periods": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schedule.PeriodInput"
                    }
                }
            }
        },
        "schedule.Regime": {
            "type": "object",
            "required": [
                "createdAt",
                "effectiveDay",
                "effectiveMonth",
                "id",
                "name",
                "periods"
            ],
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "effectiveDay": {
                    "type": "integer"
                },
                "effectiveMonth": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "periods": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schedule.Period"
                    }
                }
            }
        },
        "schedule.RegimeInput": {
            "type": "object",
            "required": [
                "effectiveDay",
                "effectiveMonth",
                "name"
            ],
            "properties": {
                "effectiveDay": {
                    "type": "integer"
                },
                "effectiveMonth": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "\"JWT Bearer token; prefix with 'Bearer '.\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}